Logic Dictionaries can now define custom macros that will be added reguardless of if they exist in logic
Started the base code for dynamic options

Randomized item entry for location can now take invalid items allowing for custom inputs

Added a reference dict for locations

Reference dicts now hold 2 copies of items and location with one of them being in quotes. This allows macros and items/locations with the same name to exist with the single quotes denoting it's not a macro

Other misc cleanup.

added support for logic entries detailing the need for x amoount of an item. example: (itemHeart, 4)