Entrances are mostly functional, they still need so cleaning up UI wise and need a playthrough test (Once I get OOTR Done).
OOTR support is mostly finished, just a few edge cases I'm not sure how to handle yet.

Macros can now be given display names, these will only be used if the macro is a trick and is displayed in the trick toggle list.

Fixed a bug where if a location is checked that contains an item that unlocks an unradomized check containing the same item, the unrandomized location would not uncheck its self. This was because the check will now unlock it's self with it's own unrandomized item. To fix this, all unrandomized item/entrances are unchecked and their items unaquired before logic calculation starts, this should probably be done for macros as well (Like the StrictLogicHandeling option did in the old tracker), but doing this already causes slow down in logic calculation.

Fixed a bug where options entries in logic wouldn't parse correctly if a macro shared the same name as the option.

If a check only has a single valid item that can be placed at it, automatically assign that item as the randomized item when it's being checked.

Fixed the UI updating multiple times for every option check in a listbox. Updates only happen once now.

Fixed toggle options not checking themseleves in the dropdown menu if the values weren't formatted a certain way.