Fixed a bug with entrance entries being imported as both entrances and Macros. I have no idea how this entrances even functioned before.
GetLocationEntryType can now properly detect exits as the exit type.

Cleaned up Requirement and conditional check

The tracker will now check paired entries when an exit is checked. It will only check the exit if it's available.

Fixed exit names when the randomized destination is being displayed in a listbox. to show "Region <= From" instead of "Region <= Region".

Fixed availability checks in the Show logic UI