Added the ability to define Variables in the dictionary that can be used in tracker in various ways.
These variables will eventually be made editable to allow for far more complex logic entries.

There are 4 types of valid Variable:

String:
String Variables can be used in option and "Item at check" checks in place of values. For example
we set the variable "mask" with the value: "MaskZora" and the following logic entry "ItemNotebook == mask"
would return true if the check "ItemNotebook" contained "MaskZora"

Bool:
Bool Variables can be used in logic as a replacement for simple true/false checks. If we create a variable "Accessible" and set it's value as "false" the logic entry "Accessible" would always return false in the logic calculation

Int:
The int variable can be used in the X amount of Item logic entries. For exxample "ItemBombchu, 5" which returns true if 5 of "ItemBombchu" have been obtained can instead be written as "ItemBombchu, intVar". This would return true if the value if "intVar" is greater than or equal to the number of "ItemBombchu" obtained

List:
List variables can be used instead of items in checks. If you create a variable called "HeartPieces" with a list of all heart pieces items. "HeartPieces" will return true if ay of those items have been obtained. This can be combined with required amounts, for example "HeartPieces, 12" will return true if 12 heart pieces have been obtained.