List Variables can now be used as both the option and the value of an option logic entry. If it's used as an option it will check if any of the options in the list match the value and if its used as a value, it will check if any value in the vist matches the option.
List Variables are now less restrictive when used as a logic item. The items in the list no longer have to be items as the full requirement check function is run on each entry, allowing them to be any type of valid logic entry, even another list variable.

String Variables can now be used standalone in logic, the LogicEntryAquired will simply be run on the strings value.