Finished OOTR support. Bugs surely exist and my debuggin code is most likely not alterting me for all logic entries that aren't being sucessfully parsed, but it should be ready enough for a test playthrough.
Variables are now displayed in the options menu and in listboxes if the setting is enabled. Variables can be marked "Static" to be lieft out of any list.

The OptionEntry check will now compare against the values of a variable if the variable is not a list.

Added a new Form for assigning variables. It can show either a numbericUpDown or a text box, which covers every variable type except bool, which can just be toggled when clicked. Lists are enter via a coma seperated list in the text box. This form is now also used to assign text to hint checks.

Fixed some bugs in the logic display, the goto macro windows will now properly only show macros and the comma seperating conditional items has been changed to a pipe to help distinguish logic names with commas in the (Specificaly the "x of item" entries)