Applied a number of optimizations to speed up logic calculation.
Options were reformatted to make better use dictionaries for logic replacements. The code that actually replaces the logic has been cleaned up as well. This was by far the biggest improvement to performance.

Logic is now only obtained once at the beggining of logic calculation. Previously it was obtained multiple times a cycle for macro and unrandomized item calculations.

In the show logic form, the "Area reached" check is now properly applied to exits.