Cleaned up the get logic function.
Getting the original logic is no longer a seperate funtion, it is now and option in the getlogic function

Logic overrides are now considered the "original" logic. Most of the use cases for logic override is assigning static macros logic so this made the most sense.

Completely removed the Dynamic logic code. Everything it could do logic options can do cleaner.