OOTR Bridge and Ganon key (item) amounts are now set via the spoiler log.
Gift from sages original item is now marked as ganons boss key. It is never listed in the spoiler, so it will be set to unrandomized and given Ganons boss key which is the only item it can be. If Ganons boss key is not on that check, it is forced junk.