Fixed OOTR Wallet Logic Names
The "Ganon" check will now always be kept randomized when importing a spoiler log but will be assigned its original items. Simply so you have something to check to signify beating the game.

OOTR Entrances now properly unrandomize when they are absent from teh spoiler log.

Added code to handle unrandomized entrances similarly to how unrandomized items are handeled.

Fixed a lot of weirdness with how hints were displayed

The price of a check is now displayed when it is checked or marked.

The set check price funtion now uses the VariableInputWindow instead of a VB input box.

The RabdomuizedStateEditor form can now toggle to show entrances and can properly change their randomized state.

Added a form to change options. Current options consist of "Check for updates (non functional)", "Horizontal Layout" and "Show Tooltips" as well as the ability to change form font.

Options can now be save to %appdata$ as default settings that are loaded automatcially when an isntance is created.