Added the OOTR Dictionary to the trackers Dictionary directory.
Added formatting to the entrance list.

the start and end combo boxes will now keep their value through UI updates.

The swap button now properly swaps the start and end combo boxes.

Added a dev tool that will simply export the names of all locations hints and entrances that are checked. This allows a consistance method of backing up the tracker if major changes are made that break save files.

Added a context menu option for area headers to "Navigate to this area" and "Filter by this area".

Added a font example to the Option screen.

Areas are now listed alphabetically in the pathfinder combo boxes.

Added an option to OOTR that allows for filtering checks at different ages or show both. This is very useful for pathfinding.