Price data is now properly assinged to Macros when importing an MMR spoiler log
Removed the Macro Dictionary options for "Static" and "Requirement/Conditional Overrides". Static macros can now be created by adding an entry for them in "additional logic" in the dictionary. Requirement and conditional ovrrides can be done by creating a static macro via the methods above and replacing the macro entry in logic with the new entry using an option. This allows for more flexability using the option entries whitelist and balcklist.