Cleaned up a lot of the PMR code
Sort of fixed the inconsistency with starting partners. The option to have them usable at all times is also added and effects logic. Unfortunately it doesn't look like the spoiler log give any idication of what partner you start with, so this will always be manual.

Fixed a loop with partner logic caused by a replace.

Added the code to get seed settings from the web based on the seed number of the spoiler log file. Doesn't do anything now but is a good start for spoiler log support.