Finished the rewrite of Paper Mario Rando. Everything looks to be functioning correctly but testing is needed.
Overhauled how Variable arrays work a bit.
If an item exists in a variable array and the array is referenced in a multi item logic entry (ex: [arrayvar, 6] ) each instance of that item that has been obtained will count. If you want an item to only count once, it can be written as a count entry in the variable (ex: [item, 1]).