Started Work on the Spoiler log Lookup Form. Generatoing a playthrough is currently the only thing working.
Changed the way Unlock data Works to fix the playthrough generator. Previously unlock data would be constantly updated as new items were unlocked which meant the playthrough generator would not always use the first set of conditionals that unlocked a check. For the playthrough generator we always want to use the first method obtained to simulate Gameplay. Unlock data is now only written when the check is unlocked and not again unless the check becomes unavailable and then available again.