The playthrough generator will now properly handle missing or invalid WinCons. If no wincon is selected it will generate a full playthrough reguardless of the Important only check box. If the WinCon is unobtainable, an error will display and no playthrough will be generated.
Added a spoiler log lookup. All Checks with assigned items (Spoiler data, manaully assigned items or Vanilla items for unrandomized checks) will be displayed in the list and will display the area, phere or check they are located in. Items with the same name are grouped together and selecting them will give a list of their location/are/sphere.