The "Root" area which serves as the starting area for entrance rando can now be defined in the logic dictionary. If it is not present it will default to "Root".
The Win Condition of a seed can now be defined in the logic dictionary. the defined item will automatiically be set as the selected object for the Playthrough generator. If it's an item, it will also be added to the required items in the seed checker.

Additional logic will now override logic in the logic file if it shares the same ID.