Add a proper function to get the name of a location entry mirroring the function to do the same for items.
Adjusted some code to properly account for the fact that exit entries do not need the required area to be a requirement in logic since logic calculation will manually check for access to the area the exit exists in.

Logic display will now automatically add the area to the requirements of an exit entry if it doesn't exist already.