Implimented basic Spoiler log importing for the OOTMM Combo Rando
Added the ability for the "Price" property to carry more data. Currently it can carry a price and a currency marker, but nothing is set to use it at the moment. This is in an attempt to hopefully solve the issue of the Combo rando having two seperate wallets that unlock different prices for different checks. Also probably broke something obscure that will take me hours to figure out later.

Renamed "Logic Options" to "Item Pool Options" in the Windows form Interface.